# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedBook - Doctor Appointment Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPaymentStatus } from "./bookingPaymentStatus";
import type { BookingSessionType } from "./bookingSessionType";
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  patientId?: number;
  sessionId: number;
  tokenNumber: number;
  status: BookingStatus;
  paymentStatus: BookingPaymentStatus;
  amountPaid?: number;
  doctorName?: string;
  hospitalName?: string;
  specialty?: string;
  date?: Date;
  sessionType?: BookingSessionType;
  createdAt?: Date;
}