# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedBook - Doctor Appointment Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  sessionId: number;
}